+++ kora-app-annotation-processor/src/test/java/ru/tinkoff/kora/kora/app/annotation/processor/WrappedDependenciesTests.java
@@ -7,6 +7,63 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class WrappedDependenciesTests extends AbstractKoraAppTest {
 
     @Test
+    public void testWrappedDependencyWithClass() {
+        var draw = compile("""
+            @KoraApp
+            public interface ExampleApplication {
+            
+                @Root
+                default Class2 class2(Class1 class1) {
+                    return new Class2();
+                }
+
+                @Root
+                default Class3 class3(ValueOf<Class1> class1) {
+                    return new Class3();
+                }
+
+                @Root
+                default Class4 class4(All<ValueOf<Class1>> class1) {
+                    return new Class4();
+                }
+
+                @Root
+                default Class2 class2ValueWrapped(Wrapped<Class1> class1) {
+                    return new Class2();
+                }
+
+                @Root
+                default Class3 class3Wrapped(ValueOf<Wrapped<Class1>> class1) {
+                    return new Class3();
+                }
+
+                @Root
+                default Class4 class4Wrapped(All<ValueOf<Wrapped<Class1>>> class1) {
+                    return new Class4();
+                }
+
+
+                default Wrapped<Class1> class1() {
+                    var c1 = new Class1();
+                    return () -> c1;
+                }
+
+                class Class1 {}
+
+                class Class2 {}
+
+                class Class3 {}
+
+                class Class4 {}
+            }
+            """);
+
+        assertThat(draw.getNodes()).hasSize(10);
+        var materializedGraph = draw.init();
+        assertThat(materializedGraph).isNotNull();
+    }
+
+    @Test
     public void testWrappedDependencyWithGeneric1() {
         var draw = compile("""
             @KoraApp
